Move Model/States serialization tests into their own test file
